refactor(web/hooks): reorganize hooks and improve chat persistence functionality
- Moved `useChatPersistence` and `useThreads` hooks to a dedicated hooks directory for better structure and maintainability.
- Updated import paths in components to reflect the new hook locations.
- Introduced `useChatPersistence` to manage chat thread loading, saving, and title updating, enhancing message persistence capabilities.
- Added `useThreads` hook for managing user threads, including fetching and deleting threads, improving overall thread management.
